Basic start to using double precision  - I have hard coded the type it uses and it only works with floats atm  - Need to implement kernels  - need to sort out GUI  - need to look into saving data as 64-bi  - maybe more?

+++ src/gui/controls/imageplot/imageplotwidget.h
@@ -215,10 +215,10 @@ private:
 
     ShowComplex complex_type;
 
-    std::vector<std::complex<float>> data_complex; // only used when we have a complex image
+    std::vector<std::complex<double>> data_complex; // only used when we have a complex image
 
-    std::vector<float> calculateComplexData() {
-        std::vector<float> im_d(data_complex.size());
+    std::vector<double> calculateComplexData() {
+        std::vector<double> im_d(data_complex.size());
 
         if (complex_type == ShowComplex::Real) {
             for (int i = 0; i < data_complex.size(); ++i)
