+++ src/simulation/cif/supercell.cpp
@@ -6,16 +6,16 @@
 
 namespace CIF {
 
-    void makeSuperCell(CIFReader cif, SuperCellInfo info, std::vector<std::string> &A, std::vector<float> &x,
-                       std::vector<float> &y, std::vector<float> &z, std::vector<float> &occ, std::vector<float> &ux,
-                       std::vector<float> &uy, std::vector<float> &uz) {
+    void makeSuperCell(CIFReader cif, SuperCellInfo info, std::vector<std::string> &A, std::vector<double> &x,
+                       std::vector<double> &y, std::vector<double> &z, std::vector<double> &occ, std::vector<double> &ux,
+                       std::vector<double> &uy, std::vector<double> &uz) {
         makeSuperCell(cif, info.uvw, info.abc, info.widths, info.tilts, A, x, y, z, occ, ux, uy, uz);
     }
 
     void makeSuperCell(CIFReader cif, Eigen::Vector3d uvw, Eigen::Vector3d abc, Eigen::Vector3d widths,
-                       Eigen::Vector3d tilts, std::vector<std::string> &A, std::vector<float> &x, std::vector<float> &y,
-                       std::vector<float> &z, std::vector<float> &occ, std::vector<float> &ux, std::vector<float> &uy,
-                       std::vector<float> &uz) {
+                       Eigen::Vector3d tilts, std::vector<std::string> &A, std::vector<double> &x, std::vector<double> &y,
+                       std::vector<double> &z, std::vector<double> &occ, std::vector<double> &ux, std::vector<double> &uy,
+                       std::vector<double> &uz) {
         // TODO: check that the uvw and abc vectors are no colinear
         UnitCell cell = cif.getUnitCell();
 
@@ -215,7 +215,7 @@ namespace CIF {
         }
     }
 
-    bool testInRange(Eigen::Vector3d pos, float xmin, float xmax, float ymin, float ymax, float zmin, float zmax) {
+    bool testInRange(Eigen::Vector3d pos, double xmin, double xmax, double ymin, double ymax, double zmin, double zmax) {
         return pos(0) >= xmin && pos(0) <= xmax && pos(1) >= ymin && pos(1) <= ymax && pos(2) >= zmin && pos(2) <= zmax;
     }
