+++ src/simulation/structure/atom.h
@@ -5,35 +5,35 @@
 struct Atom
 {
     unsigned int A;
-    float x, y, z;
+    double x, y, z;
 
-    Atom() : A(0), x(0.0f), y(0.0f), z(0.0f) {}
+    Atom() : A(0), x(0.0), y(0.0), z(0.0) {}
 
-    Atom(unsigned int _A, float _x, float _y, float _z) : A(_A), x(_x), y(_y), z(_z) {}
+    Atom(unsigned int _A, double _x, double _y, double _z) : A(_A), x(_x), y(_y), z(_z) {}
 
     // this is only for position testing (Not element testing)
     bool operator==(const Atom &RHS) {
         return x == RHS.x && y == RHS.y && z == RHS.z;
     }
 
-    Atom operator* (float f) {
+    Atom operator* (double f) {
         return {A, f*x, f*y, f*z};
     }
 };
 
 struct AtomSite : public Atom
 {
-    float occ;
-    float ux, uy, uz;
-    AtomSite() : Atom(), occ(1.0f), ux(0.0f), uy(0.0f), uz(0.0f) {}
-    AtomSite(unsigned int _A, float _x, float _y, float _z, float _occ, float _ux, float _uy, float _uz) : Atom(_A, _x, _y, _z), occ(_occ), ux(_ux), uy(_uy), uz(_uz) {}
+    double occ;
+    double ux, uy, uz;
+    AtomSite() : Atom(), occ(1.0), ux(0.0), uy(0.0), uz(0.0) {}
+    AtomSite(unsigned int _A, double _x, double _y, double _z, double _occ, double _ux, double _uy, double _uz) : Atom(_A, _x, _y, _z), occ(_occ), ux(_ux), uy(_uy), uz(_uz) {}
 
-    AtomSite operator* (float f) {
+    AtomSite operator* (double f) {
         // u is in Angstrom squared
         return AtomSite(A, f*x, f*y, f*z, occ, f*f*ux, f*f*uy, f*f*uz);
     }
 
-    void setThermal(float u) {
+    void setThermal(double u) {
         ux = u;
         uy = u;
         uz = u;
