+++ src/simulation/structure/thermalvibrations.cpp
@@ -9,11 +9,11 @@ ThermalVibrations::ThermalVibrations() {
     force_defined = false;
     u_default = 0.0f;
     std::vector<int> set_elements = {};
-    std::vector<float> u_squareds = std::vector<float>(Utils::VectorSymbolToNumber.size(), u_default);
+    std::vector<double> u_squareds = std::vector<double>(Utils::VectorSymbolToNumber.size(), u_default);
 }
 
-std::vector<float> ThermalVibrations::getDefinedVibrations() {
-    std::vector<float> vals;
+std::vector<double> ThermalVibrations::getDefinedVibrations() {
+    std::vector<double> vals;
     for (const auto& i : set_elements) {
         vals.emplace_back(getVibrations(i)); // -1 as hydrogen is 1, but element 0, etc...
     }
@@ -21,14 +21,14 @@ std::vector<float> ThermalVibrations::getDefinedVibrations() {
     return vals;
 }
 
-void ThermalVibrations::setVibrations(float def, std::vector<int> elements, std::vector<float> vibs) {
+void ThermalVibrations::setVibrations(double def, std::vector<int> elements, std::vector<double> vibs) {
     if (elements.size() != vibs.size())
         throw(std::runtime_error("Cannot set thermal vibrations with different elements and vibration vector sizes"));
 
     u_default = def;
     // reset to our default, check we have the vector is need be
     if (u_squareds.empty())
-        u_squareds = std::vector<float>(Utils::VectorSymbolToNumber.size(), u_default);
+        u_squareds = std::vector<double>(Utils::VectorSymbolToNumber.size(), u_default);
     std::fill(u_squareds.begin(), u_squareds.end(), u_default);
 
     // fill in the non default values
@@ -41,9 +41,9 @@ void ThermalVibrations::setVibrations(float def, std::vector<int> elements, std:
     set_elements = elements;
 }
 
-float ThermalVibrations::getVibrations(unsigned int element) {
+double ThermalVibrations::getVibrations(unsigned int element) {
     if (u_squareds.empty())
-        u_squareds = std::vector<float>(Utils::VectorSymbolToNumber.size(), u_default);
+        u_squareds = std::vector<double>(Utils::VectorSymbolToNumber.size(), u_default);
 
     if (element >= u_squareds.size())
         throw(std::runtime_error("Cannot access thermal parameters for element: " + std::to_string(element)));
